+++ packages/core/src/data/ome_zarr/metadata_loaders.ts
@@ -1,12 +1,12 @@
 import * as zarr from "zarrita";
-import { Image as ImageV04 } from "./ome_zarr/0.4/image";
-import { Plate as PlateV04 } from "./ome_zarr/0.4/plate";
-import { Well as WellV04 } from "./ome_zarr/0.4/well";
-import { Image } from "./ome_zarr/0.5/image";
-import { Plate } from "./ome_zarr/0.5/plate";
-import { Well } from "./ome_zarr/0.5/well";
-import { OmeZarrImageSource } from "./ome_zarr_image_source";
-import { Version as ZarrVersion, openGroup } from "./zarr/open";
+import { Image as ImageV04 } from "./0.4/image";
+import { Plate as PlateV04 } from "./0.4/plate";
+import { Well as WellV04 } from "./0.4/well";
+import { Image } from "./0.5/image";
+import { Plate } from "./0.5/plate";
+import { Well } from "./0.5/well";
+import { OmeZarrImageSource } from "./image_source";
+import { Version as ZarrVersion, openGroup } from "../zarr/open";
 
 const versions = ["0.4", "0.5"] as const;
 const versionsSet: ReadonlySet<string> = new Set(versions);
